monitoreo/requirements.txt
Django==4.2.7
Pillow==10.0.1
openpyxl==3.1.2
python-decouple==3.8
mysqlclient==2.2.0
gunicorn==21.2.0
whitenoise==6.6.0